Move rightmost failure tracking to its own module
Extract all rightmost failure state and functions from trace.ts into
lib/rightmostFailure.ts. This also removes the inlined line table logic
since the new module can import directly from position.ts.
