+++ lib/index.ts
@@ -4,3 +4,4 @@ export * from "./trace.js";
 export * from "./types.js";
 export * from "./tarsecError.js";
 export * from "./position.js";
+export * from "./rightmostFailure.js";
